fix: move bricks to collision layer 2
Bricks were in the same collision layer as walls. This caused bricks
which were touching walls to immediately collide. This solution moves
bricks into collision layer 2 and also adds the ball to the same layer
(ball is in collision layers 1 and 2). An alternative solution would be
to check if a collision with the brick is from a ball, but that may be a
performance hit since bricks which are touching walls may collide
repeatedly.

Resolves #1